requirements.txt
altair==4.2.0
altgraph==0.17
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.3
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-rest-knox==4.2.0
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
Flask==2.0.1
fonttools==4.28.5
future==0.18.2
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.26
idna==2.10
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pefile==2021.5.13
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
pure-eval==0.2.1
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
Pympler==1.0.1
pyowm==3.2.0
pyparsing==3.0.6
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
pyzmq==22.3.0
regex==2021.11.10
requests==2.25.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
sqlparse==0.4.1
stack-data==0.1.4
streamlit==1.4.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.4
validators==0.18.2
virtualenv==20.13.0
vk-api==11.9.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
